% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentiationCholesky.R
\name{ADchol}
\alias{ADchol}
\title{construct object for Automated Differentiation Cholesky decomposition}
\usage{
ADchol(P_list)
}
\arguments{
\item{P_list}{a list of symmetric matrices of class spam, each of dimension \eqn{q \times q},
and with sum of the matrices assumed to be positive definite.}
}
\value{
An object of class `ADchol`.
}
\description{
Construct object for reverse Automated Differentiation of Cholesky decomposition,
with as input a list of semi-positive symmetric sparse matrices \eqn{P_i}, each of
dimension \eqn{q \times q}. The function `ADchol` calculates the matrix \eqn{C}, the sum
the precision matrices \eqn{P_i}: \eqn{C = \sum_{i}  P_i}. Next, it calculates the Cholesky
Decomposition using the multiple minimum degree (MMD) algorithm
of the `spam` package.
}
\keyword{internal}
