% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APSIMdat}
\alias{APSIMdat}
\title{Simulated Biomass as function of time using APSIM wheat.}
\format{
A data.frame with 121 rows and 4 columns.
\describe{
\item{env}{Environment, Emerald in 1993}
\item{geno}{Simulated genotype g001}
\item{das}{Days after sowing}
\item{biomass}{Simulated biomass using APSIM; medium measurement error added}
}
}
\usage{
APSIMdat
}
\description{
Simulated Biomass as function of time using APSIM wheat.
}
\references{
Bustos-Korts et al. (2019) Combining Crop Growth Modeling and
Statistical Genetic Modeling to Evaluate Phenotyping Strategies
\doi{10.3389/FPLS.2019.01491}
}
\keyword{datasets}
