% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley.uniformity.trial}
\alias{barley.uniformity.trial}
\title{Uniformity trial of barley}
\format{
A data.frame with 1076 rows and 3 columns
\describe{
 \item{row}{row coordinate}
 \item{col}{column coordinate}
 \item{yield}{yield per plot}
}
}
\source{
H. P. Piepho & E. R. Williams (2010). Linear variance models
for plant breeding trials. Plant Breeding, 129, 1-8.
\doi{10.1111/j.1439-0523.2009.01654.x}
}
\usage{
barley.uniformity.trial
}
\description{
Uniformity trial of barley
}
\references{
Piepho, Hans‐Peter, Martin P. Boer, and Emlyn R. Williams.
"Two‐dimensional P‐spline smoothing for spatial analysis of plant breeding trials."
Biometrical Journal 64, no. 5 (2022): 835-857.
}
\keyword{datasets}
