% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinesFunctions.R
\name{Bsplines}
\alias{Bsplines}
\title{Construct design matrix for B-Splines}
\usage{
Bsplines(knots, x, deriv = 0)
}
\arguments{
\item{knots}{A numerical vector of knot positions.}

\item{x}{a numeric vector of values at which to evaluate the B-spline
functions or derivatives.}

\item{deriv}{A numerical value. The derivative of the given order is
evaluated at the x positions.}
}
\description{
Construct design matrix for B-Splines.
}
