% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFunctions.R
\name{calcStandardErrors}
\alias{calcStandardErrors}
\title{Standard errors for predictions}
\usage{
calcStandardErrors(C, D)
}
\arguments{
\item{C}{a symmetric matrix of class spam}

\item{D}{a matrix of class spam}
}
\value{
a vector with standard errors
}
\description{
Calculates the standard errors for predictions \eqn{D \hat{u}}, see Welham et al. 2004 for
details.
}
\details{
The prediction error variance is given by \eqn{\sigma^2 D C^{-1} D'},
where \eqn{C} is the mixed model coefficient matrix, and \eqn{D} defines linear combinations of fixed
and random effects, see Welham et al. 2004 for details. The standard errors are given by the the square root of
the diagonal. To calculate the standard errors in an efficient way we use that

\deqn{\frac{\partial log|C + \xi_i DD'|}{\partial \xi_i} |_{\xi_i=0}
 = trace(C^{-1} d_i d_i') =
trace(d_i' C^{-1} d_i) = d_i' C^{-1} d_i, }
where \eqn{d_i} is column \eqn{i} of matrix \eqn{D}. The values of \eqn{d_i' C^{-1} d_i} can
be calculated more efficient by using Automated Differentiation of the Choleksy algorithm, see
Smith 1995 for details.
}
\keyword{internal}
