% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcUsc}
\alias{calcUsc}
\title{Spectral decomposition of D'D}
\usage{
calcUsc(q, ord)
}
\arguments{
\item{q}{A numeric value.}

\item{ord}{A numeric value.}
}
\value{
A q x (q-ord) matrix, the spectral decomposition D'D.
}
\description{
Spectral decomposition of D'D
}
