% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkFormVars}
\alias{checkFormVars}
\title{Check variables in formula}
\usage{
checkFormVars(formula, data)
}
\description{
Check that all variables in a formula are present in the data. The variables
are converted to a factor in the data.frame as well.
}
\keyword{internal}
