% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkGroup}
\alias{checkGroup}
\title{Check group part}
\usage{
checkGroup(random, group)
}
\description{
Check that all variables in group are also specified as grp() in random part
of the model and vice versa.
}
\keyword{internal}
