% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFactor.R
\name{condFactor}
\alias{condFactor}
\title{Function to analyse cf terms in the random part.}
\usage{
condFactor(random, data)
}
\description{
Function to analyse cf terms in the random part.
}
\keyword{internal}
