% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructCCt}
\alias{constructCCt}
\title{Construct constraint matrix}
\usage{
constructCCt(q, pord)
}
\arguments{
\item{q}{dimension of the B-spline basis used.}

\item{pord}{order of the penalty matrix (pord=1 or 2).}
}
\value{
a q x q matrix of type spam
}
\description{
Construct constraint matrix
}
\keyword{internal}
