% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructPenalty}
\alias{constructPenalty}
\title{construct P-splines penalty matrix D'D}
\usage{
constructPenalty(q, pord)
}
\arguments{
\item{q}{integer with dimensions}

\item{pord}{order of the penalty}
}
\value{
qxq matrix D'D of class spam
}
\description{
construct P-splines penalty matrix D'D
}
