% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructRinv.R
\name{constructRinv}
\alias{constructRinv}
\title{Helper function for constructing Rinv}
\usage{
constructRinv(df, residual, weights)
}
\description{
Helper function for constructing Rinv
}
\keyword{internal}
