% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructX}
\alias{constructX}
\title{construct fixed part of the spline model}
\usage{
constructX(B, x, scaleX, pord)
}
\arguments{
\item{B}{matrix with B-spline basis.}

\item{x}{vector with values for x}

\item{scaleX}{logical. If scaleX is FALSE, the original x is used. If scaleX is TRUE,
scaling is used, based on the B-splines basis. For details see the code}

\item{pord}{order of the penalty, values 1 or 2.}
}
\value{
a matrix X
}
\description{
construct fixed part of the spline model
}
