% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentiationCholesky.R
\name{DerivCholesky}
\alias{DerivCholesky}
\title{This function saves result of partial derivatives of Cholesky to a
a spam matrix. It is assumed that cholC and objAD have same values, i.e.:
- cholC has been updated with:
    C <- LMMsolver:::linearSum(theta, listP)
    cholC <- update(cholC, C)
- objAD has been calculated with:
    objAD <- LMMsolver:::ADchol(listP)
    x <- LMMsolver:::dlogdet(objAD, theta)}
\usage{
DerivCholesky(cholC, objAD)
}
\description{
this function can be useful for example for calculating standard errors.
}
\keyword{internal}
