% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticsMME.R
\name{diagnosticsMME}
\alias{diagnosticsMME}
\title{Give diagnostics for mixed model coefficient matrix C and the cholesky
decomposition}
\usage{
diagnosticsMME(object)
}
\arguments{
\item{object}{an object of class LMMsolve.}
}
\value{
A summary of the mixed model coefficient matrix and its choleski
decomposition.
}
\description{
Give diagnostics for mixed model coefficient matrix C and the cholesky
decomposition
}
\examples{
## Fit model on oats data
data(oats.data)

## Fit simple model with only fixed effects.
LMM1 <- LMMsolve(fixed = yield ~ rep + gen,
                data = oats.data)

## Obtain deviance.
diagnosticsMME(LMM1)

}
