% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandGinv}
\alias{expandGinv}
\title{Combine two lists of Ginv matrices}
\usage{
expandGinv(lGinv1, lGinv2)
}
\arguments{
\item{lGinv1}{a list of sparse matrices.}

\item{lGinv2}{a list of sparse matrices.}
}
\value{
a list of sparse matrices of dimension (d1+d2) x (d1+d2)
}
\description{
output is a combined list, of dimension (d1+d2) x (d1+d2),
with matrices in lGinv1 extended to [A1 0]
                                    [0  0]
and              lGinv2 extended to [0  0]
                                    [0 A2]
}
\keyword{internal}
