% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve-class.R
\name{new_LMMsolve}
\alias{new_LMMsolve}
\title{Class LMMsolve}
\description{
Class LMMsolve of fitted mixed models.
}
\section{Slots}{

An object of class LMMsolve contains the following slots.
\describe{
\item{logL}{The restricted log-likelihood at convergence}
\item{sigma2e}{The residual error}
\item{tau2e}{The estimated variance components}
\item{ED}{The effective dimensions}
\item{EDmax}{The maximal effective dimensions}
\item{EDnames}{The names of the effective dimensions}
\item{a}{The estimated effects from the mixed model equations}
\item{yhat}{The fitted values}
\item{dim}{The dimensions for each of the fixed and random terms in the
mixed model}
\item{term.labels}{The Names of the fixed and random terms in the mixed
model}
\item{splRes}{An object with definition of spline argument}
}
}

