% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve-class.R
\name{LMMsolveObject}
\alias{LMMsolveObject}
\title{Fitted LMMsolve Object}
\value{
An object of class \code{LMMsolve} contains the following components:
\item{logL}{The restricted log-likelihood at convergence}
\item{dev}{The REML deviance at convergence (i.e., - 2 times the
restricted log-likelihood \code{logL})}
\item{sigma2e}{The residual error}
\item{tau2e}{The estimated variance components}
\item{ED}{The effective dimensions}
\item{EDmax}{The maximal effective dimensions}
\item{EDnames}{The names of the effective dimensions}
\item{a}{The estimated effects from the mixed model equations}
\item{yhat}{The fitted values}
\item{dim}{The dimensions for each of the fixed and random terms in the
mixed model}
\item{term.labels}{The Names of the fixed and random terms in the mixed
model}
\item{splRes}{An object with definition of spline argument}
}
\description{
An object of class \code{LMMsolve} returned by the LMMsolve function,
representing a fitted linear mixed model. Objects of this class have
methods for the generic functions coef, .....
}
