% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolver.R
\docType{package}
\name{LMMsolver}
\alias{LMMsolver}
\alias{LMMsolver-package}
\title{Package LMMsolver}
\description{
Linear Mixed Model Solver using sparse matrix algebra.
}
\details{
An efficient and flexible system to solve sparse mixed model
equations, for models that are often used in statistical genetics.
Important applications are the use of splines to model spatial or temporal
trends. Another application area is mixed model QTL analysis for
multiparental populations, allowing for heterogeneous residual variance and
random design matrices with Identity-By-Descent (IBD) probabilities.
}
\references{
Martin P. Boer (2023). \emph{Tensor product P-splines using a sparse mixed model formulation},
Statistical Modelling, 23, p. 465 - 479. \doi{10.1177/1471082X231178591}
}
\author{
Martin Boer \email{martin.boer@wur.nl}

Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (maintainer)
}
