% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve-class.R
\name{logLik.LMMsolve}
\alias{logLik.LMMsolve}
\title{Log-likelihood of an LMMsolve object}
\usage{
\method{logLik}{LMMsolve}(object, ...)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{...}{some methods for this generic require additional arguments.
None are used in this method.}
}
\description{
Obtain the Restricted Maximum Log-Likelihood of a model fitted using
LMMsolve.
}
