% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikelihood.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Function to obtain restricted log-likelihood given values for penalty parameter}
\usage{
logLikelihood(object, theta)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{theta}{a matrix with values of precision parameters theta.}
}
\value{
A data.frame with logL and the derivatives.
}
\description{
Function to obtain restricted log-likelihood given values for penalty parameter
}
