% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLogLik.R
\name{mLogLik}
\alias{mLogLik}
\title{Function to obtain restricted log-likelihood and the first derivatives of the log-likelihood,
given values for the penalty parameters}
\usage{
mLogLik(object, theta)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{theta}{a matrix with values of precision parameters theta.}
}
\value{
A data.frame with logL and the first derivatives of log-likelihood
}
\description{
Function to obtain restricted log-likelihood and the first derivatives of the log-likelihood,
given values for the penalty parameters
}
