% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{Multinomialsolve}
\alias{Multinomialsolve}
\title{Solver for multinomial distribution, at the moment only for 1D splines}
\usage{
Multinomialsolve(fixed, dat, family = multinomial())
}
\description{
Solver for multinomial distribution, at the moment only for 1D splines
}
