% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multipop}
\alias{multipop}
\title{Simulated QTL mapping data set}
\format{
A data.frame with 180 rows and 6 columns.
\describe{
\item{cross}{Cross ID, two populations, AxB and AxC}
\item{ind}{Genotype ID}
\item{pA}{Probability that individual has alleles from parent A}
\item{pB}{Probability that individual has alleles from parent B}
\item{pC}{Probability that individual has alleles from parent C}
\item{pheno}{simulated phenotypic value}
}
}
\usage{
multipop
}
\description{
Simulated QTL mapping data set
}
\keyword{datasets}
