% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nameCoefs}
\alias{nameCoefs}
\title{Helper function for naming coefficients.}
\usage{
nameCoefs(coefs, desMat, termLabels, s, e, data, group = NULL, type = "fixed")
}
\description{
Helper function for naming coefficients.
}
\keyword{internal}
