% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinesFunctions.R
\name{ndxMatrix}
\alias{ndxMatrix}
\title{Construct index matrix}
\usage{
ndxMatrix(df, lZ, names)
}
\arguments{
\item{df}{A data.frame.}

\item{lZ}{A list of matrices.}

\item{names}{A character vector of names.}
}
\value{
The index matrix.
}
\description{
Construct index matrix.
}
\keyword{internal}
