% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oats.data}
\alias{oats.data}
\title{Alpha lattice design of spring oats}
\format{
A data.frame with 72 rows and 7 columns
\describe{
\item{plot}{plot number}
\item{rep}{replicate}
\item{block}{incomplete block}
\item{gen}{genotype}
\item{yield}{dry matter yield}
\item{row}{row}
\item{col}{column}
}
}
\source{
J. A. John & E. R. Williams (1995). Cyclic and computer
generated designs. Chapman and Hall, London. Page 146.
}
\usage{
oats.data
}
\description{
Alpha lattice design of spring oats
}
\details{
The response is grain yield in kg per hectare. The design was an alpha design
with 24 varieties, three replicates and  six incomplete blocks of size four per replicate.
The 72 plots were arranged in a single linear array.
}
\references{
Boer, Martin P., Hans-Peter Piepho, and Emlyn R. Williams.
"Linear variance, P-splines and neighbour differences for spatial adjustment
in field trials: how are they related?." JABES 25, no. 4 (2020): 676-698.
}
\keyword{datasets}
