% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainSmoothTrend.R
\name{obtainSmoothTrend}
\alias{obtainSmoothTrend}
\title{Obtain Smooth Trend.}
\usage{
obtainSmoothTrend(
  object,
  grid = NULL,
  newdata = NULL,
  includeIntercept = FALSE
)
}
\arguments{
\item{object}{An object of class LMMsolve.}

\item{grid}{A numeric vector having the length of the dimension of the fitted
spline component. This represents the number of grid points at which a
surface will be computed.}

\item{newdata}{...}

\item{includeIntercept}{Should the value of the intercept be included in
the computed smooth trend?}
}
\value{
A data.frame with predictions for the smooth trend on the specified
grid.
}
\description{
Obtain the smooth trend for models fitted with a spline component.
}
