% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainSmoothTrend.R
\name{obtainSmoothTrend1D}
\alias{obtainSmoothTrend1D}
\title{Obtain Smooth Trend for 1D P-splines}
\usage{
obtainSmoothTrend1D(object, grid)
}
\arguments{
\item{object}{an object of class LMMsolve.}

\item{grid}{number of grid points.}
}
\description{
Obtain Smooth Trend for 1D P-splines
}
\keyword{internal}
