% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap3Dfast.R
\name{predict.sap3Dfast}
\alias{predict.sap3Dfast}
\title{predict for sap3D, without spectral decomposition}
\usage{
\method{predict}{sap3Dfast}(object, ..., newdata, grid = c(30, 30, 30))
}
\description{
predict for sap3D, without spectral decomposition
}
