% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl2D.R
\name{predict.spl2Dfast}
\alias{predict.spl2Dfast}
\title{predict for spl2D, without spectral decomposition}
\usage{
\method{predict}{spl2Dfast}(object, ..., grid)
}
\arguments{
\item{object}{...}

\item{...}{...}

\item{grid}{...}
}
\description{
predict for spl2D, without spectral decomposition
}
