% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl3D.R
\name{predict.spl3Dfast}
\alias{predict.spl3Dfast}
\title{predict for spl3D, without spectral decomposition}
\usage{
\method{predict}{spl3Dfast}(object, ..., newdata, grid = c(30, 30, 30))
}
\arguments{
\item{object}{...}

\item{...}{...}

\item{newdata}{...}

\item{grid}{...}
}
\description{
predict for spl3D, without spectral decomposition
}
