% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFunctions.R
\name{predictTest}
\alias{predictTest}
\title{Test function for predict, for the moment internal}
\usage{
predictTest(object, classify)
}
\description{
Test function for predict, for the moment internal
}
\keyword{internal}
