% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeIntercept}
\alias{removeIntercept}
\title{Remove the intercept from a design matrix}
\usage{
removeIntercept(X)
}
\arguments{
\item{X}{design matrix.}
}
\value{
a matrix if \code{X} has more than one column, otherwise return NULL
}
\description{
Remove the intercept from a design matrix
}
\keyword{internal}
