% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinesFunctions.R
\name{RowKronecker}
\alias{RowKronecker}
\title{Row-wise kronecker product}
\usage{
RowKronecker(X1, X2)
}
\arguments{
\item{X1}{A matrix.}

\item{X2}{A matrix.}
}
\value{
The row-wise kronecker product of X1 and X2.
}
\description{
Row-wise kronecker product
}
\keyword{internal}
