% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap2Dfast.R
\name{sap2D}
\alias{sap2D}
\title{sap2D, without spectral decomposition}
\usage{
sap2D(x1, x2, knots, scaleX = FALSE)
}
\arguments{
\item{x1}{...}

\item{x2}{...}

\item{knots}{...}

\item{scaleX}{...}
}
\description{
sap2D, without spectral decomposition
}
