% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap3Dfast.R
\name{sap3D}
\alias{sap3D}
\title{sap3D, without spectral decomposition}
\usage{
sap3D(x1, x2, x3, knots, scaleX = TRUE)
}
\arguments{
\item{x1}{...}

\item{x2}{...}

\item{x3}{...}

\item{knots}{...}

\item{scaleX}{...}
}
\description{
sap3D, without spectral decomposition
}
