% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap3Dfast.R
\name{sap3Dfast}
\alias{sap3Dfast}
\title{sap3D, without spectral decomposition}
\usage{
sap3Dfast(y, x1, x2, x3, knots, trace = TRUE, tolerance = 1e-08, scaleX = TRUE)
}
\arguments{
\item{y}{...}

\item{x1}{...}

\item{x2}{...}

\item{x3}{...}

\item{knots}{...}

\item{trace}{...}

\item{tolerance}{...}

\item{scaleX}{...}
}
\description{
sap3D, without spectral decomposition
}
