% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SeaSurfaceTemp}
\alias{SeaSurfaceTemp}
\title{Sea Surface Temperature}
\format{
A data.frame with 15607 rows and 4 columns.
\describe{
\item{lon}{longitude}
\item{lat}{latitude}
\item{sst}{sea surface temperature in Kelvin}
\item{type}{defines training and test set}
}
}
\usage{
SeaSurfaceTemp
}
\description{
Sea Surface Temperature
}
\references{
Cressie et al. (2022) Basis-function models in spatial statistics.
Annual Review of Statistics and Its Application.
\doi{10.1146/annurev-statistics-040120-020733}
}
\keyword{datasets}
