% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMixedModels.R
\name{sparseMixedModels}
\alias{sparseMixedModels}
\title{importFrom stats update}
\usage{
sparseMixedModels(
  y,
  X,
  Z,
  lGinv,
  lRinv,
  maxit = 100,
  tolerance = 1e-06,
  trace = FALSE,
  theta = NULL
)
}
\description{
importFrom stats update
}
\keyword{internal}
