% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl1D.R
\name{spl1D}
\alias{spl1D}
\title{fit 1D P-splines}
\usage{
spl1D(x, nseg, pord=2, degree=3, scaleX=TRUE, xlim=NULL)
}
\arguments{
\item{x}{numerical vector containing the values of x covariate.}

\item{nseg}{number of segments}

\item{pord}{order of penalty, default \code{pord=2}}

\item{degree}{degree of B-spline basis, default \code{degree=3}}

\item{scaleX}{logical, scale fixed effect or not. Default is TRUE,
no scaling.}

\item{xlim}{numerical vector of length 2 containing the domain of covariate
x where the knots should be placed. Default set to \code{NULL} (covariate range).}
}
\value{
A list with the following elements:
\itemize{
  \item \code{X} - design matrix for fixed effect. The intercept is not included.
  \item \code{Z} - design matrix for random effect.
  \item \code{lGinv} - a list of precision matrices
  \item \code{knots} - a list of vectors with knot positions
}
}
\description{
fit 1D P-splines using sparse implementation.
}
