% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap2D.R
\name{spl2Dfast}
\alias{spl2Dfast}
\title{spl2D, without spectral decomposition}
\usage{
spl2Dfast(y, x1, x2, nseg, trace = TRUE, tolerance = 1e-08, scaleX = FALSE)
}
\arguments{
\item{y}{...}

\item{x1}{numerical vector containing the values of \code{x1} covariate.}

\item{x2}{numerical vector containing the values of \code{x2} covariate.}

\item{nseg}{number of segments}

\item{trace}{Should the progress of the algorithm be printed? Default \code{trace=FALSE}.}

\item{tolerance}{A numerical value. The convergence tolerance for the
modified Henderson algorithm to estimate the variance components.}

\item{scaleX}{logical, scale fixed effect or not. Default is TRUE,
no scaling.}
}
\description{
spl2D, without spectral decomposition
}
