% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap3D.R
\name{spl3Dfast}
\alias{spl3Dfast}
\title{spl3D, without spectral decomposition}
\usage{
spl3Dfast(y, x1, x2, x3, nseg, trace = TRUE, tolerance = 1e-08, scaleX = TRUE)
}
\arguments{
\item{y}{...}

\item{x1}{numerical vector containing the values of \code{x1} covariate.}

\item{x2}{numerical vector containing the values of \code{x2} covariate.}

\item{x3}{numerical vector containing the values of \code{x3} covariate.}

\item{nseg}{number of segments}

\item{trace}{Should the progress of the algorithm be printed? Default \code{trace=FALSE}.}

\item{tolerance}{A numerical value. The convergence tolerance for the
modified Henderson algorithm to estimate the variance components.}

\item{scaleX}{logical, scale fixed effect or not. Default is TRUE,
no scaling.}
}
\description{
spl3D, without spectral decomposition
}
