% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentiationCholesky.R
\name{ADchol}
\alias{ADchol}
\title{Automatic differentiation Cholesky, ZtZ and P spam matrices,
with C = lambda[1]*P1 + lambda[2]*P2 + .....}
\usage{
ADchol(P_list)
}
\description{
Automatic differentiation Cholesky, ZtZ and P spam matrices,
with C = lambda[1]*P1 + lambda[2]*P2 + .....
}
