% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve.R
\name{coef.LMMsolve}
\alias{coef.LMMsolve}
\title{Obtain the coefficients from the mixed model equations}
\usage{
\method{coef}{LMMsolve}(object)
}
\arguments{
\item{object}{an object of class LMMsolve}
}
\description{
Obtain the coefficients from the mixed model equations
}
