% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructGinvSplines}
\alias{constructGinvSplines}
\title{Construct list of Ginv matrices of splines}
\usage{
constructGinvSplines(q, knots, pord)
}
\arguments{
\item{q}{vector, with dimension of the B-spline basis used.}

\item{knots}{list with knot positions for each dimension}

\item{pord}{order of the penalty matrix (1 or 2).}
}
\value{
a list of symmetric matrices of length of vector q.
}
\description{
Construct list of Ginv matrices of splines
}
\keyword{internal}
