% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructPenalty}
\alias{constructPenalty}
\title{Construct P-splines penalty matrix D'D}
\usage{
constructPenalty(q, pord, dx)
}
\arguments{
\item{q}{integer with dimensions.}

\item{pord}{order of the penalty.}

\item{dx}{distance between the knots.}
}
\value{
qxq matrix D'D of class spam
}
\description{
Construct P-splines penalty matrix D'D
}
\keyword{internal}
