% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructPenalty}
\alias{constructPenalty}
\title{Construct the penalty matrix}
\usage{
constructPenalty(q, pord, dx)
}
\arguments{
\item{q}{integer with dimensions.}

\item{pord}{order of the penalty.}

\item{dx}{distance between the knots.}
}
\value{
qxq penalty matrix of class spam
}
\description{
Construct a scaled version of the P-splines penalty matrix, see details.
}
\details{
The P-spline penalty matrix has the form \eqn{D'D}, where \eqn{D} is the `pord` order
difference matrix. To make the penalty matrix more stable if there are many knots,
a scaled version is used, \eqn{(1/dx)^(2pord-1) D'D}.
}
\keyword{internal}
