% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentiationCholesky.R
\name{DerivCholesky}
\alias{DerivCholesky}
\title{This function saves result of partial derivatives of Cholesky to a
a spam matrix, and is used to calculate standard errors and for predictions.}
\usage{
DerivCholesky(cholC)
}
\description{
This function saves result of partial derivatives of Cholesky to a
a spam matrix, and is used to calculate standard errors and for predictions.
}
\keyword{internal}
