% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve-class.R
\name{deviance.LMMsolve}
\alias{deviance.LMMsolve}
\title{Deviance of an LMMsolve object}
\usage{
\method{deviance}{LMMsolve}(object, ...)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{...}{some methods for this generic require additional arguments.
None are used in this method.}
}
\description{
Obtain the deviance of a model fitted using LMMsolve.
}
