% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve-class.R
\name{diagnosticsMME}
\alias{diagnosticsMME}
\title{Give diagnostics for mixed model coefficient matrix C and the cholesky
decomposition}
\usage{
diagnosticsMME(object)
}
\arguments{
\item{object}{an object of class LMMsolve}
}
\description{
Give diagnostics for mixed model coefficient matrix C and the cholesky
decomposition
}
