% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve.R
\name{displayMME}
\alias{displayMME}
\title{Display the sparseness of the mixed model coefficient matrix.}
\usage{
displayMME(object, cholesky = FALSE)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{cholesky}{logical. If \code{cholesky = TRUE} it will plot the cholesky.}
}
\description{
Display the sparseness of the mixed model coefficient matrix.
}
