% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlogdet}
\alias{dlogdet}
\title{Calculate the partial derivatives of log-determinant.}
\usage{
dlogdet(arg, theta)
}
\arguments{
\item{arg}{argument of class ADchol.}

\item{theta}{a vector with precision parameters}
}
\value{
The gradient with partial derivatives of \eqn{log|C|} with respect to
parameters \eqn{\theta_i}. As attribute `logdet` \eqn{log|C|} is returned.
}
\description{
This function calculates the partial derivatives of the the log-determinant in an
efficient way, by using reverse Automated Differentiation of the Cholesky Algorithm,
see Smith (1995) for details. Let
 \deqn{C = \sum_{i} \theta_i P_i}
where the matrices \eqn{P_i} are stored in the `ADchol` object. The partial derivatives
of matrix \eqn{C} are defined by:
\deqn{\frac{\partial C}{\partial \theta_i} = \text{trace} [C^{-1} P_i]},
but are calculated more efficient using backwards Automated Differentiation.
}
\references{
Smith, S. P. (1995). Differentiation of the Cholesky algorithm.
Journal of Computational and Graphical Statistics, 4(2), 134-147.
}
\keyword{internal}
