% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve.R
\name{LMMsolve}
\alias{LMMsolve}
\title{Solve Linear Mixed Models}
\usage{
LMMsolve(
  fixed,
  random = NULL,
  spatial = NULL,
  group = NULL,
  lGinverse = NULL,
  data,
  residual = NULL,
  tolerance = 1e-06,
  trace = FALSE,
  display = FALSE,
  maxit = 250
)
}
\arguments{
\item{fixed}{A formula for the fixed part of the model. Should be of the
form "response ~ pred"}

\item{random}{A formula for the random part of the model. Should be of the
form "~ pred".}

\item{group}{A named list where each component is a numeric vector
specifying contiguous fields in data that are to be considered as a
single term.}

\item{lGinverse}{A named list where each component is a matrix corresponding
to the group of the same name.}

\item{data}{A data.frame containing the modeling data.}

\item{residual}{A formula for the residual part of the model. Should be of
the form "~ pred".}

\item{tolerance}{A numerical value. The convergence tolerance for the
algorithm.}

\item{trace}{Should the progress of the algorithm be printed?}

\item{display}{Should the ... matrix created in the algorithm be plotted?}

\item{maxit}{A numerical value. The maximum number of iterations for the
algorithm.}

\item{spatialA}{formula for the random part of the model. Should be of the
form "~ sap2D()" or "~sap3d()".}
}
\value{
An object of class LMMsolve, a list with the following items:
\item{logL}{The loglikelihood}
\item{sigma2e}{The residual error}
\item{tau2e}{}
\item{ED}{The effective dimensions}
\item{EDmax}{The maximal effective dimensions}
\item{EDnames}{The names of the effective dimensions}
\item{a}{}
\item{yhat}{The fitted values}
\item{dim}{}
\item{term.labels}{}
}
\description{
Solve Linear Mixed Models
}
