% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve.R
\name{LMMsolve}
\alias{LMMsolve}
\title{Solve Linear Mixed Models}
\usage{
LMMsolve(
  fixed,
  random = NULL,
  spline = NULL,
  group = NULL,
  data,
  residual = NULL,
  tolerance = 1e-06,
  trace = FALSE,
  maxit = 250,
  omitConstant = TRUE
)
}
\arguments{
\item{fixed}{A formula for the fixed part of the model. Should be of the
form "response ~ pred"}

\item{random}{A formula for the random part of the model. Should be of the
form "~ pred".}

\item{spline}{A formula for the spline part of the model. Should be of the
form "~ spl1D()", ~ spl2D()" or "~spl3D()".}

\item{group}{A named list where each component is a numeric vector
specifying contiguous fields in data that are to be considered as a
single term.}

\item{data}{A data.frame containing the modeling data.}

\item{residual}{A formula for the residual part of the model. Should be of
the form "~ pred".}

\item{tolerance}{A numerical value. The convergence tolerance for the
modified Henderson algorithm to estimate the variance components.}

\item{trace}{Should the progress of the algorithm be printed? Default
\code{trace = FALSE}.}

\item{maxit}{A numerical value. The maximum number of iterations for the
algorithm. Default \code{maxit = 250}.}

\item{omitConstant}{Omit the constant in the restricted log-likelihood. Default is
\code{TRUE}, as for example in \code{asreml}. In \code{nlme} and SAS the constant
is included.}
}
\value{
An object of class \code{LMMsolve} representing the fitted model.
See \code{\link{LMMsolveObject}} for a full description of the components in
this object.
}
\description{
Solve Linear Mixed Models.
}
\seealso{
\code{\link{LMMsolveObject}}
}
