% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve.R
\name{LMMsolve}
\alias{LMMsolve}
\title{Solve Linear Mixed Models}
\usage{
LMMsolve(
  fixed,
  random = NULL,
  group = NULL,
  lGinverse = NULL,
  data,
  residual = NULL,
  tolerance = 1e-06,
  trace = FALSE,
  display = FALSE,
  maxit = 250
)
}
\arguments{
\item{fixed}{...}

\item{random}{...}

\item{group}{...}

\item{lGinverse}{...}

\item{data}{...}

\item{residual}{...}

\item{tolerance}{...}

\item{trace}{...}

\item{display}{...}

\item{maxit}{...}
}
\description{
Solve Linear Mixed Models
}
