% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spl1D.R
\name{obtainSmoothTrend1D}
\alias{obtainSmoothTrend1D}
\title{obtain Smooth Trend for 1D P-splines}
\usage{
obtainSmoothTrend1D(object, grid)
}
\arguments{
\item{grid}{number of grid points.}

\item{obj}{an object of class LMMsolve}
}
\description{
obtain Smooth Trend for 1D P-splines
}
