% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainSmoothTrend.R
\name{obtainSmoothTrend2D}
\alias{obtainSmoothTrend2D}
\title{Obtain Smooth Trend for 2D P-splines.}
\usage{
obtainSmoothTrend2D(object, grid)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{grid}{a numeric vector of length 2, with the number of grid points at
which a two-dimensional surface will be computed.}
}
\description{
Obtain Smooth Trend for 2D P-splines.
}
\keyword{internal}
