% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainSmoothTrend.R
\name{obtainSmoothTrend3D}
\alias{obtainSmoothTrend3D}
\title{Obtain Smooth Trend for 3D P-splines}
\usage{
obtainSmoothTrend3D(object, grid)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{grid}{a numeric vector of length 3, with the number of grid points at
which a three-dimensional surface will be computed.}
}
\description{
Obtain Smooth Trend for 3D P-splines
}
\keyword{internal}
