% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMsolve-class.R
\name{predict.LMMsolve}
\alias{predict.LMMsolve}
\title{Predict function}
\usage{
\method{predict}{LMMsolve}(object, newdata, se.fit = FALSE)
}
\arguments{
\item{object}{an object of class LMMsolve.}

\item{newdata}{A data.frame containing new points for which the smooth
trend should be computed. Column names should include the names used when
fitting the spline model.}

\item{se.fit}{calculate standard errors, default \code{FALSE}.}
}
\value{
A data.frame with predictions for the smooth trend on the specified
grid. The standard errors are saved if `se.fit=TRUE`.
}
\description{
Predict function
}
