% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap2D.R
\name{predict.sap2Dfast}
\alias{predict.sap2Dfast}
\title{predict for sap2D, without spectral decomposition}
\usage{
\method{predict}{sap2Dfast}(object, ..., grid)
}
\arguments{
\item{object}{...}

\item{...}{...}

\item{grid}{...}
}
\description{
predict for sap2D, without spectral decomposition
}
