% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap2D.R
\name{sap2D}
\alias{sap2D}
\title{fit 2D P-splines}
\usage{
sap2D(x1, x2, nseg, pord=2, degree=3, scaleX=FALSE, x1lim=NULL,x2lim=NULL)
}
\arguments{
\item{x1}{numerical vector containing the values of \code{x1} covariate.}

\item{x2}{numerical vector containing the values of \code{x2} covariate.}

\item{nseg}{number of segments}

\item{pord}{order of penalty, default \code{pord=2}}

\item{degree}{degree of B-spline basis, default \code{degree=3}}

\item{scaleX}{logical, scale fixed effect or not. Default is FALSE,
no scaling.}

\item{x1lim}{numerical vector of length 2 containing the domain of covariate
\code{x1} where the knots should be placed. Default set to \code{NULL} (covariate range).}

\item{x2lim}{numerical vector of length 2 containing the domain of covariate
\code{x2} where the knots should be placed. Default set to \code{NULL} (covariate range).}
}
\value{
A list with three matrices: design matrix for fixed effect \code{X} (not including the
intercept), the design matrix for random effect \code{Z}, and a list of precision matrices \code{Ginv}$ for
the random effect \code{Z}.
}
\description{
fit 2D P-splines using sparse implementation.
}
