% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap2D.R
\name{sap2Dfast}
\alias{sap2Dfast}
\title{sap2D, without spectral decomposition}
\usage{
sap2Dfast(y, x1, x2, knots, trace = TRUE, tolerance = 1e-08, scaleX = FALSE)
}
\arguments{
\item{y}{...}

\item{x1}{numerical vector containing the values of \code{x1} covariate.}

\item{x2}{numerical vector containing the values of \code{x2} covariate.}

\item{trace}{Should the progress of the algorithm be printed?}

\item{tolerance}{A numerical value. The convergence tolerance for the
algorithm.}

\item{scaleX}{logical, scale fixed effect or not. Default is FALSE,
no scaling.}
}
\description{
sap2D, without spectral decomposition
}
