% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap2Dfast.R
\name{sap2Dfast}
\alias{sap2Dfast}
\title{sap2D, without spectral decomposition}
\usage{
sap2Dfast(y, x1, x2, knots, trace = TRUE, thr = 1e-08, scaleX = FALSE)
}
\description{
sap2D, without spectral decomposition
}
