% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap3D.R
\name{sap3D}
\alias{sap3D}
\title{fit 3D P-splines}
\usage{
sap3D(
  x1,
  x2,
  x3,
  nseg,
  pord = 2,
  degree = 3,
  scaleX = TRUE,
  x1lim = NULL,
  x2lim = NULL,
  x3lim = NULL
)
}
\arguments{
\item{x1}{numerical vector containing the values of \code{x1} covariate.}

\item{x2}{numerical vector containing the values of \code{x2} covariate.}

\item{x3}{numerical vector containing the values of \code{x3} covariate.}

\item{nseg}{number of segments}

\item{pord}{order of penalty, default \code{pord=2}}

\item{degree}{degree of B-spline basis, default \code{degree=3}}

\item{scaleX}{logical, scale fixed effect or not. Default is TRUE,
no scaling.}

\item{x1lim}{numerical vector of length 2 containing the domain of covariate
\code{x1} where the knots should be placed. Default set to \code{NULL} (covariate range).}

\item{x2lim}{numerical vector of length 2 containing the domain of covariate
\code{x2} where the knots should be placed. Default set to \code{NULL} (covariate range).}

\item{x3lim}{numerical vector of length 2 containing the domain of covariate
\code{x3} where the knots should be placed. Default set to \code{NULL} (covariate range).}
}
\value{
A list with the following elements:
\itemize{
  \item \code{X} - design matrix for fixed effect. The intercept is not included.
  \item \code{Z} - design matrix for random effect.
  \item \code{lGinv} - a list of precision matrices
  \item \code{knots} - a list of vectors with knot positions
}
}
\description{
fit 3D P-splines using sparse implementation.
}
