% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap3Dfast.R
\name{sap3Dfast}
\alias{sap3Dfast}
\title{sap3D, without spectral decomposition}
\usage{
sap3Dfast(y, x1, x2, x3, knots, trace = TRUE, thr = 1e-08, scaleX = TRUE)
}
\description{
sap3D, without spectral decomposition
}
