% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLMMsolve.R
\name{summary.LMMsolve}
\alias{summary.LMMsolve}
\title{summary function}
\usage{
\method{summary}{LMMsolve}(object, which = "dimensions", ...)
}
\arguments{
\item{object}{an object of class LMMsolve}

\item{which}{default dimensions (only option so far).}

\item{\dots}{some methods for this generic require additional arguments.
None are used in this method.}
}
\description{
Prints the effective dimensions of the model.
}
